requirements.txt
asgiref==3.3.4
Django==3.2.3
django-suit @ https://github.com/darklow/django-suit/tarball/v2
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
asgiref==3.3.4
Django==3.2.3
django-suit==2.0a1
mysql-connector-python==8.0.25
Pillow==8.2.0
protobuf==3.17.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
